+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed `Villain` Archetype to memorise spells like an Arcanist instead of Wizard. The Spell slots progression is the same as the spells per day progression.
 - Changed `Body Strengthening` to give DR/— per level.
 - Update `Spell Master` feature icon.
+- Renamed `Protective Aura` to `Aura of Golden Protection`.
+- Renamed `Glorious Aura` to `Aura of Majesty`.
 ### Removed
 - Removed `Harem Magnet` ability.
 - Removed `True Main Character` feature.